[Fix] Rename the component Timer/ShowTime
